% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mesh.2.5D}
\alias{plot.mesh.2.5D}
\title{Plot a mesh.2.5D object}
\usage{
\method{plot}{mesh.2.5D}(x, ...)
}
\arguments{
\item{x}{A \code{mesh.2.5D} object generated by \code{create.mesh.2.5D} or \code{second.order.mesh.2.5D}.}

\item{...}{Arguments representing graphical options to be passed to \link[graphics]{par}.}
}
\description{
Plot the triangulation of a mesh \code{mesh.2.5D} object, generated by \code{create.mesh.2.5D} or \code{second.order.mesh.2.5D}. The points indicates the mesh nodes.
}
\examples{
##Load a mesh.2.5D object
data(candy)
##Plot the triangulation of the object
plot.mesh.2.5D(candy)
}
